Move components feature and polish UI
