+++ app/src/main/kotlin/com/d4rk/android/apps/apptoolkit/app/components/ui/contract/ComponentsUnlockEvent.kt
@@ -1,4 +1,4 @@
-package com.d4rk.android.apps.apptoolkit.components.ui.contract
+package com.d4rk.android.apps.apptoolkit.app.components.ui.contract
 
 import com.d4rk.android.libs.apptoolkit.core.ui.base.handling.UiEvent
