+++ app/src/main/kotlin/com/d4rk/android/apps/apptoolkit/app/components/ui/navigation/ComponentsNavigationBuilder.kt
@@ -1,9 +1,9 @@
-package com.d4rk.android.apps.apptoolkit.components.ui.navigation
+package com.d4rk.android.apps.apptoolkit.app.components.ui.navigation
 
 import com.d4rk.android.apps.apptoolkit.app.main.ui.views.navigation.AppNavigationEntryContext
 import com.d4rk.android.apps.apptoolkit.app.main.utils.constants.AppNavKey
 import com.d4rk.android.apps.apptoolkit.app.main.utils.constants.ComponentsRoute
-import com.d4rk.android.apps.apptoolkit.components.ui.views.ComponentsRoute as ComponentsScreenRoute
+import com.d4rk.android.apps.apptoolkit.app.components.ui.views.ComponentsRoute as ComponentsScreenRoute
 import com.d4rk.android.libs.apptoolkit.core.ui.navigation.NavigationEntryBuilder
 
 /**
