+++ app/src/main/kotlin/com/d4rk/android/apps/apptoolkit/app/settings/about/ui/AppAboutSettingsContent.kt
@@ -3,8 +3,8 @@ package com.d4rk.android.apps.apptoolkit.app.settings.about.ui
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
-import com.d4rk.android.apps.apptoolkit.components.ui.ComponentsUnlockViewModel
-import com.d4rk.android.apps.apptoolkit.components.ui.contract.ComponentsUnlockEvent
+import com.d4rk.android.apps.apptoolkit.app.components.ui.ComponentsUnlockViewModel
+import com.d4rk.android.apps.apptoolkit.app.components.ui.contract.ComponentsUnlockEvent
 import com.d4rk.android.libs.apptoolkit.app.about.ui.AboutSettingsList
 import org.koin.compose.viewmodel.koinViewModel
