+++ apptoolkit/src/main/kotlin/com/d4rk/android/libs/apptoolkit/core/ui/views/fields/DatePickerTextField.kt
@@ -67,6 +67,7 @@ fun DatePickerTextField(
                 hapticFeedback.performHapticFeedback(HapticFeedbackType.ContextClick)
                 showDialog.value = true
             },
+        shape = MaterialTheme.shapes.large,
         trailingIcon = {
             Icon(Icons.Default.CalendarToday, contentDescription = null)
         },
